- SunPro 5.3 workaround.

+++ Packages/Number_types/src/MP_Float.C
@@ -296,7 +296,8 @@ to_double_exp(const MP_Float &b)
     d += d_exp * b.of_exp(i);
   }
 
-  return std::make_pair(d, exp * log_limb);
+  // the (int) cast is necessary for SunPro 5.3...
+  return std::make_pair(d, (int) exp * log_limb);
 }
 }
